mini_twitter/api/urls/__init__.py
from django.urls import path, include
from .author import author_router
from .post import post_router
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path("", include(author_router.urls)),
    path("", include(post_router.urls)),
    path('login/', TokenObtainPairView.as_view(), name='login'),
    path('refresh/', TokenRefreshView.as_view(), name='refresh')
]